# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/buildings/{buildingId}/floors/{floorId}/devices": {
      "get": {
        "operationId": "DevicesController_getFloorDevices",
        "parameters": [
          {
            "name": "buildingId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "E",
              "type": "string"
            }
          },
          {
            "name": "floorId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devices catalogue returned successfully"
          },
          "404": {
            "description": "Floor not found"
          }
        },
        "summary": "Get catalogue and effective marker positions for a floor",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/iot/buildings/{buildingId}/floors/{floorId}/devices": {
      "get": {
        "operationId": "DevicesController_getIotFloorDevices",
        "parameters": [
          {
            "name": "buildingId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "E",
              "type": "string"
            }
          },
          {
            "name": "floorId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "IoT floor devices returned successfully"
          }
        },
        "summary": "Get IoT devices list directly from upstream API (Phase 06, no persistence)",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/iot/devices/{deviceId}/telemetry": {
      "get": {
        "operationId": "DevicesController_getDeviceTelemetry",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "d0ba7a2a-f40d-4b39-885b-6dc79815a104",
              "type": "string"
            }
          },
          {
            "name": "start",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stop",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Telemetry readings returned successfully"
          },
          "400": {
            "description": "Invalid query parameters (range, limit)"
          },
          "404": {
            "description": "Device not found in registry"
          }
        },
        "summary": "Get IoT device telemetry readings/events within rolling 72-hour window (Phase 07, zero persistence)",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/devices/{deviceId}": {
      "get": {
        "operationId": "DevicesController_getDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "d0ba7a2a-f40d-4b39-885b-6dc79815a104",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device details returned successfully"
          },
          "404": {
            "description": "Device not found"
          }
        },
        "summary": "Get device metadata and placement detail",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/devices/{deviceId}/display-position": {
      "put": {
        "operationId": "DevicesController_updateDisplayPosition",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-expected-placement-revision",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Expected-Placement-Revision",
            "in": "header",
            "description": "Current known revision number to prevent stale writes",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDisplayPositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position override saved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: device display positions cannot be modified"
          },
          "409": {
            "description": "Revision conflict or floor context changed"
          }
        },
        "summary": "Create or update display position override",
        "tags": [
          "Devices"
        ]
      },
      "delete": {
        "operationId": "DevicesController_resetDisplayPosition",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-expected-placement-revision",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "buildingId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "floorId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Expected-Placement-Revision",
            "in": "header",
            "description": "Current known revision number to prevent stale writes",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Position override reset successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: device display positions cannot be modified"
          },
          "409": {
            "description": "Revision conflict"
          }
        },
        "summary": "Reset display position override back to original",
        "tags": [
          "Devices"
        ]
      }
    },
...
    "/api/v1/dashboard/buildings/{buildingId}/iot/devices": {
      "get": {
        "description": "Returns a normalized active device catalogue for Dashboard Page 07. Supports full-building catalogue (omitting floorId) and approved floor-scoped queries (floorId=4, 6). Zero database persistence.",
        "operationId": "DashboardIotController_getCatalogue",
        "parameters": [
          {
            "name": "buildingId",
            "required": true,
            "in": "path",
            "description": "Building identifier (currently E only)",
            "schema": {
              "example": "E",
              "type": "string"
            }
          },
          {
            "name": "floorId",
            "required": false,
            "in": "query",
            "description": "Application floor ID (4, 6). When omitted, returns full unfiltered catalogue.",
            "schema": {
              "example": "4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device catalogue returned successfully with live provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDeviceCatalogueResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid buildingId, unsupported floor, or floor G"
          },
          "502": {
            "description": "Upstream IoT service error, timeout, or malformed payload"
          },
          "503": {
            "description": "IoT integration disabled, fixture mode, or missing credentials"
          }
        },
        "summary": "Get Dashboard IoT device catalogue (read-only, live provenance)",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "GIS-UIT Building E Digital Twin - Backend API",
    "description": "Internal API for device catalogue, marker positions, and persistence in Building E.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "Coordinate3DDto": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "example": 1.5
          },
          "y": {
            "type": "number",
            "example": 0
          },
          "z": {
            "type": "number",
            "example": -2
          }
        },
        "required": [
          "x",
          "y",
          "z"
        ]
      },
      "UpdateDisplayPositionDto": {
        "type": "object",
        "properties": {
          "buildingId": {
            "type": "string",
            "example": "E"
          },
          "floorId": {
            "type": "string",
            "example": "4"
          },
          "frameId": {
            "type": "string",
            "example": "E/4/floor-local"
          },
          "frameVersion": {
            "type": "number",
            "example": 1
          },
          "position": {
            "$ref": "#/components/schemas/Coordinate3DDto"
          }
        },
        "required": [
          "buildingId",
          "floorId",
          "frameId",
          "frameVersion",
          "position"
        ]
      },
...
      "DashboardCatalogueProvenanceDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "example": "live",
            "enum": [
              "live"
            ]
          },
          "sourceId": {
            "type": "string",
            "example": "iot-device-catalogue"
          },
          "sourceType": {
            "type": "string",
            "example": "iot_backend"
          },
          "fetchedAt": {
            "type": "string",
            "example": "2026-09-26T02:45:00.000Z",
            "description": "Application backend fetch timestamp (ISO-8601)"
          },
          "caveats": {
            "example": [
              "Dữ liệu development fallback được áp dụng cho tầng 4"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "mode",
          "sourceId",
          "sourceType",
          "fetchedAt"
        ]
      },
      "DashboardCatalogueMappingDto": {
        "type": "object",
        "properties": {
          "floorMode": {
            "type": "string",
            "example": "TEST_CURRENT_FLOOR_4_6_V1"
          },
          "developmentFallbackApplied": {
            "type": "boolean",
            "example": false,
            "description": "True when development fallback from floor 0 was applied"
          },
          "requestedUpstreamFloorLevel": {
            "type": "number",
            "example": 4,
            "nullable": true,
            "description": "Mapped upstream floor level requested, or null if full catalogue"
          }
        },
        "required": [
          "floorMode",
          "developmentFallbackApplied",
          "requestedUpstreamFloorLevel"
        ]
      },
      "DashboardCatalogueSummaryDto": {
        "type": "object",
        "properties": {
          "receivedCount": {
            "type": "number",
            "example": 25,
            "description": "Raw records received from upstream"
          },
          "acceptedCount": {
            "type": "number",
            "example": 25,
            "description": "Valid records accepted into catalogue"
          },
          "skippedCount": {
            "type": "number",
            "example": 0,
            "description": "Malformed records skipped"
          },
          "duplicateCount": {
            "type": "number",
            "example": 0,
            "description": "Duplicate device IDs skipped"
          },
          "truncated": {
            "type": "boolean",
            "example": false,
            "nullable": true,
            "description": "Upstream truncation indicator"
          }
        },
        "required": [
          "receivedCount",
          "acceptedCount",
          "skippedCount",
          "duplicateCount",
          "truncated"
        ]
      },
      "DashboardDeviceSourceLocationDto": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "example": 10.5,
            "description": "Source installation X coordinate"
          },
          "y": {
            "type": "number",
            "example": 2.1,
            "description": "Source installation Y coordinate"
          },
          "z": {
            "type": "number",
            "example": 15.3,
            "description": "Source installation Z coordinate"
          },
          "floorLevel": {
            "type": "number",
            "example": 4,
            "description": "Source installation floor level integer"
          }
        },
        "required": [
          "x",
          "y",
          "z",
          "floorLevel"
        ]
      },
      "DashboardDeviceCatalogueItemDto": {
        "type": "object",
        "properties": {
          "externalDeviceId": {
            "type": "string",
            "example": "70B3D57ED0073E9D",
            "description": "Opaque upstream device identifier"
          },
          "sourceDeviceType": {
            "type": "string",
            "example": "solar",
            "description": "Raw source device type string"
          },
          "category": {
            "type": "string",
            "example": "solar",
            "description": "Mapped application device category"
          },
          "active": {
            "type": "boolean",
            "example": true,
            "description": "Catalogue registration/activity metadata (NOT an online/health signal)"
          },
          "sourceCreatedAt": {
            "type": "string",
            "example": "2026-09-20T19:09:17.355Z",
            "description": "Source record creation timestamp"
          },
          "sourceUpdatedAt": {
            "type": "string",
            "example": "2026-09-20T19:09:17.355Z",
            "description": "Source metadata update timestamp (NOT telemetry last-seen)"
          },
          "sourceLocation": {
            "$ref": "#/components/schemas/DashboardDeviceSourceLocationDto"
          },
          "displayFloorId": {
            "type": "string",
            "example": "4",
            "nullable": true,
            "description": "Viewer display floor ID, or null if unmapped"
          },
          "floorAssignment": {
            "type": "string",
            "example": "source",
            "enum": [
              "source",
              "development-fallback",
              "unmapped"
            ],
            "description": "Floor assignment provenance: source, development-fallback, or unmapped"
          }
        },
        "required": [
          "externalDeviceId",
          "sourceDeviceType",
          "category",
          "active",
          "sourceCreatedAt",
          "sourceUpdatedAt",
          "sourceLocation",
          "displayFloorId",
          "floorAssignment"
        ]
      },
      "DashboardDeviceCatalogueResponseDto": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "example": 1
          },
          "buildingId": {
            "type": "string",
            "example": "E"
          },
          "requestedFloorId": {
            "type": "string",
            "example": "4",
            "nullable": true
          },
          "availability": {
            "type": "string",
            "example": "ready",
            "enum": [
              "ready",
              "empty"
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/DashboardCatalogueProvenanceDto"
          },
          "mapping": {
            "$ref": "#/components/schemas/DashboardCatalogueMappingDto"
          },
          "summary": {
            "$ref": "#/components/schemas/DashboardCatalogueSummaryDto"
          },
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardDeviceCatalogueItemDto"
            }
          }
        },
        "required": [
          "schemaVersion",
          "buildingId",
          "requestedFloorId",
          "availability",
          "provenance",
          "mapping",
          "summary",
          "devices"
        ]
      },
...
    }
  }
}